make person class and teacher class
